BaseController and base Script
Base script can be changed to add functionality specific to the base it is controlling. I suggest using this as a parent class and making an Army/Farm base class which inherits this. The BaseContoller script will create a list of all objects of type Base and when asked to will activate a random one. In the future we can modify this to sort into an Army and Farm list for better control, and add other sorting functionality as well. 
These classes allow the base objects to be placed wherever we want them and will not interfere with that.